# docs/api/openapi.yaml
...
openapi: '3.0.3'
...
info:
  title: 'CIE v2.3.2 - Catalog Intelligence Engine'
  version: '2.3.2'
  description: |
    Pre-publish enforcement gates, semantic validation, AI citation audit,
    and tier management for 120+ staff e-commerce operations.

    ## Architecture
    This spec covers TWO distinct API layers:
     - tag group 'Engine' = Python Engine endpoints (called BY the PHP CMS)
     - tag group 'CMS' = PHP CMS endpoints (called BY the frontend UI)
    All validation is server-side. The PHP CMS never bypasses the Python Engine.
    Auth: internal bearer tokens per service. No public access.

# SOURCE: INSTRUCTION 2 — cie_v231_openapi.yaml servers block
servers:
  - url: https://cie.internal.example.com/api/v1
    description: Production
  - url: https://cie-staging.internal.example.com/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local Dev — Python Engine
  - url: http://localhost:8080/api/v1
    description: Local Dev — PHP CMS

# SOURCE: INSTRUCTION 3 — cie_v231_openapi.yaml top-level security block
security:
  - BearerAuth: []

# SOURCE: INSTRUCTION 4 — CIE_v2.3.1_Enforcement_Dev_Spec.pdf Section 7.1 (engine tags);
#         CIE_v232_UI_Restructure_Instructions.docx Section 4 (CMS tags)
...
paths:
...
  /sku/embed:
    # SOURCE: cie_v231_openapi.yaml /sku/{sku_id}/embed — note: path is /sku/embed per INSTRUCTION 5 list
    post:
      summary: Generate embedding vector for text
      tags: [Semantic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  maxLength: 5000
                  description: Description text to embed
      responses:
        '200':
          description: Embedding vector result
          content:
            application/json:
              schema:
                type: object
                properties:
                  vector:
                    type: array
                    items:
                      type: number
                  dimensions:
                    type: integer
                    example: 1536
                  model:
                    type: string
                    example: text-embedding-3-small
  /sku/similarity:
    # SOURCE: cie_v231_openapi.yaml /sku/{sku_id}/similarity — path as listed in INSTRUCTION 5
    post:
      summary: Compute cosine similarity vs cluster intent vector
      description: |
        Threshold: 0.72. Below threshold = description rejected.
      tags: [Semantic]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, cluster_id]
              properties:
                description:
                  type: string
                cluster_id:
                  type: string
                  example: CLU-CABLE-PENDANT-E27
      responses:
        '200':
          description: Similarity computation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarityResponse'
...
components:
  securitySchemes:
    # SOURCE: cie_v231_openapi.yaml components/securitySchemes
    BearerAuth:
      type: http
      scheme: bearer
      description: Internal service token
  schemas:
...
    SimilarityResponse:
      # SOURCE: CLAUDE.md §11 — cosine_similarity and threshold are INTERNAL ONLY.
      # Removed from wire response. PHP caller (G4_VectorGate.php) uses status + message only.
      type: object
      properties:
        status:
          type: string
          enum: [pass, fail]
        message:
          type: string
          nullable: true
...